商品

+++ src/main/java/com/example/demo/biz/impl/CatalogService.java
@@ -1,6 +1,7 @@
 package com.example.demo.biz.impl;
 
 import com.example.demo.biz.ICatalogService;
+import com.example.demo.biz.IGoodService;
 import com.example.demo.dao.ICatalogDao;
 import com.example.demo.dao.ICatalogRepository;
 import com.example.demo.dao.ISecondRepository;
@@ -24,6 +25,8 @@ public class CatalogService implements ICatalogService {
     ISecondRepository iSecondRepository;
     @Autowired
     ICatalogDao iCatalogDao;
+    @Autowired
+    IGoodService iGoodService;
     @Override
     public List<FirstlevelEntity> findAll() {
         return iCatalogRepository.findAll();
@@ -60,7 +63,9 @@ public class CatalogService implements ICatalogService {
         String[] idss=ids.split(",");
         if (idss.length>0) {
             for (String id : idss){
-                iSecondRepository.delete(Long.valueOf(id));
+                if (null!=iGoodService.selectBySecondId(Long.valueOf(id)) ){
+                    iSecondRepository.delete(Long.valueOf(id));
+                }
                 }
         }
     }
